Rename page for cart as new standard
